Fix projection methods in dials.search_beam_position
- All three projection methods now write optimised.expt file.

- Beam position is now computed for each detector in the experiment.

- per_image=True option now averages all the beam positions from
  individual images and saves it in the optimised.expt.

- Removed options for data slicing (image and imageset ranges)
  since they were not compatible with data averaging
  per detector.
