+++ src/dials/algorithms/beam_position/compute_beam_position.py
@@ -8,7 +8,7 @@ from dials.algorithms.beam_position import (
 from dials.algorithms.beam_position.plot import Figure
 
 
-def compute_beam_position(image, params, image_index=None, imageset_index=None):
+def compute_beam_position(image, params, image_index=None, detector_index=None):
     method_x, method_y = resolve_projection_methods(params)
 
     if method_x == "midpoint":
@@ -33,13 +33,13 @@ def compute_beam_position(image, params, image_index=None, imageset_index=None):
     else:
         image_str = ""
 
-    if imageset_index is not None:
-        imageset_str = "_imageset_%05d" % imageset_index
+    if detector_index is not None:
+        detector_str = "_detector_%d" % detector_index
     else:
-        imageset_str = ""
+        detector_str = ""
 
     if params.projection.plot:
-        fig = Figure(f"beam_position{imageset_str}{image_str}.png")
+        fig = Figure(f"beam_position{detector_str}{image_str}.png")
 
         fig.plot_main(image, params, beam_position=(x, y))
         solver_x.plot(fig)
@@ -47,8 +47,6 @@ def compute_beam_position(image, params, image_index=None, imageset_index=None):
 
         fig.save_and_close()
 
-    if not params.projection.per_image:
-        print(f"{x:.2f}, {y:.2f}")
     return x, y
